empty_impl_runtime_apis_call
