+++ core/sr-api-macros/tests/ui/empty_impl_runtime_apis_call.rs
@@ -1,0 +1,20 @@
+use runtime_primitives::traits::GetNodeBlockType;
+use test_client::runtime::Block;
+use client::{decl_runtime_apis, impl_runtime_apis};
+
+/// The declaration of the `Runtime` type and the implementation of the `GetNodeBlockType`
+/// trait are done by the `construct_runtime!` macro in a real runtime.
+struct Runtime {}
+impl GetNodeBlockType for Runtime {
+	type NodeBlock = Block;
+}
+
+decl_runtime_apis! {
+	pub trait Api {
+		fn test(data: u64);
+	}
+}
+
+impl_runtime_apis! {}
+
+fn main() {}
